Meme Generator/requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.24.4
pandas==2.0.3
Pillow==10.1.0
python-dateutil==2.8.2
python-docx==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
